refactor: Convert camelCase filenames to kebab-case
- Rename 22 files from camelCase to kebab-case format
- Rename testUtils/ directory to test-utils/
- Update all import statements across codebase (37 references)
- Update documentation examples in README.md
- Maintain backward compatibility for public API exports

Files renamed:
- restClient.ts → rest-client.ts
- TodoistApi.ts → todoist-api.ts
- testUtils/ → test-utils/
- urlHelpers.ts → url-helpers.ts
- multipartUpload.ts → multipart-upload.ts
- All TodoistApi test files to kebab-case format
