+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './TodoistApi'
+export * from './todoist-api'
 export * from './authentication'
 export * from './types'
 export * from './utils'
