+++ src/test-utils/mocks.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from 'axios'
-import * as restClient from '../restClient'
+import * as restClient from '../rest-client'
 
 export function setupRestClientMock(responseData: unknown, status = 200): jest.SpyInstance {
     const response = { status, data: responseData } as AxiosResponse
