+++ src/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './colors'
 export * from './sanitization'
-export { getTaskUrl, getProjectUrl, getSectionUrl } from './urlHelpers'
+export { getTaskUrl, getProjectUrl, getSectionUrl } from './url-helpers'
